requirements_.txt
amqp==2.6.1
asgiref==3.2.10
backcall==0.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-admin-list-filter-dropdown==1.0.3
django-allauth==0.42.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-database-size==1.0.2
django-debug-toolbar==2.2
django-filter==2.3.0
django-material==1.7.0
django-smuggler==1.0.1
django-timezone-field==4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
feedparser==5.2.1
fuzzywuzzy==0.18.0
gunicorn==20.0.4
hercules==0.0
idna==2.10
kombu==4.6.11
nmmd==0.0.3
oauthlib==3.1.0
psycopg2==2.8.5
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-Levenshtein==0.12.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
termcolor==1.1.0
uni==0.1
urllib3==1.25.10
vine==1.3.0
